del function and frontend css

+++ src/main/java/sakila/project/Repository/cityRepository.java
@@ -3,7 +3,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import sakila.project.entities.City;
 
-public interface cityRepository extends CrudRepository<City, Long> {
+public interface cityRepository extends CrudRepository<City, Short> {
 
     @Query(value = "SELECT * FROM city WHERE UPPER(city) = ? AND country_id = ?", nativeQuery = true)
     City SearchCity(String city_name, Short country_id);
